chore: adhere to standards + simple refactor
