+++ src/components/Feedback.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './styles/App.scss';
+import './styles/Feedback.scss';
 
 import Base from './shared/Base';
 import { HeaderSections } from './shared/PlaynetConstants';
@@ -8,7 +8,7 @@ import { HeaderSections } from './shared/PlaynetConstants';
 function Feedback(): JSX.Element {
   return (
     <div>
-      <Base section={HeaderSections.feedback}>
+      <Base section={HeaderSections.FEEDBACK}>
         <h1>Feedback</h1>
       </Base>
     </div>
