+++ src/components/shared/Base/Header.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import '../styles/App.scss';
+import '../../styles/Base.scss';
 
 import { Link } from 'react-router-dom';
-import { HeaderSections } from '../shared/PlaynetConstants';
+import { HeaderSections } from '../PlaynetConstants';
 
 
-type HeaderProps = {
+export interface HeaderProps {
   section: HeaderSections,
 }
 function Header(props: HeaderProps): JSX.Element {
@@ -15,19 +15,19 @@ function Header(props: HeaderProps): JSX.Element {
       <nav>
         <div
           id="textSpan"
-          style={(props.section === HeaderSections.intro) ? { fontWeight: 'bold' } : { fontWeight: 'normal' }}
+          style={(props.section === HeaderSections.INTRO) ? { fontWeight: 'bold' } : { fontWeight: 'normal' }}
         >
           <Link to="/">Home</Link>
         </div>
         <div
           id="textSpan"
-          style={(props.section === HeaderSections.activities) ? { fontWeight: 'bold' } : { fontWeight: 'normal' }}
+          style={(props.section === HeaderSections.ACTIVITIES) ? { fontWeight: 'bold' } : { fontWeight: 'normal' }}
         >
           <Link to="/activities">Activities</Link>
         </div>
         <div
           id="textSpan"
-          style={(props.section === HeaderSections.feedback) ? { fontWeight: 'bold' } : { fontWeight: 'normal' }}
+          style={(props.section === HeaderSections.FEEDBACK) ? { fontWeight: 'bold' } : { fontWeight: 'normal' }}
         >
           <Link to="/feedback">Feedback</Link>
         </div>
