+++ src/components/styles/Activities.scss
@@ -1,0 +1,3 @@
+div {
+  color: #000;
+}
