+++ src/components/styles/Activity1.scss
@@ -1,0 +1,3 @@
+div {
+  color: #000;
+}
