+++ src/components/styles/Activity2.scss
@@ -1,0 +1,3 @@
+div {
+  color: #000;
+}
