+++ src/components/styles/Activity3.scss
@@ -1,0 +1,3 @@
+div {
+  color: #000;
+}
