+++ src/components/styles/Feedback.scss
@@ -1,0 +1,3 @@
+div {
+  color: #000;
+}
