+++ src/components/styles/Home.scss
@@ -1,0 +1,3 @@
+div {
+  color: #000;
+}
