+++ src/index.tsx
@@ -1,36 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-} from 'react-router-dom';
-
-import Activities from './components/Activities';
-import Feedback from './components/Feedback';
-import Home from './components/Home';
+import App from './components/App';
 
 import './global.scss';
 
 ReactDOM.render(
   <React.StrictMode>
-    <Router>
-      <div>
-        {/* A <Switch> looks through its children Routes and
-            renders the first one that matches the current URL. */}
-        <Switch>
-          <Route path="/activities">
-            <Activities />
-          </Route>
-          <Route path="/feedback">
-            <Feedback />
-          </Route>
-          <Route path="/">
-            <Home />
-          </Route>
-        </Switch>
-      </div>
-    </Router>
+    <App />
   </React.StrictMode>,
   document.getElementById('root'),
 );
